+++ packages/desktop/app/components/Sidebar/SidebarWithPrograms.js
@@ -5,7 +5,11 @@ import { Sidebar } from './Sidebar';
 
 // TODO fetch programs from api
 const DUMMY_PROGRAM = { name: 'All programs', id: 'all-programs' };
-const fetchPrograms = async () => [DUMMY_PROGRAM];
+const ACTIVE_COVID_19_PROGRAM = {
+  name: 'Active COVID-19 patients',
+  id: 'active-covid-19-program',
+};
+const fetchPrograms = async () => [DUMMY_PROGRAM, ACTIVE_COVID_19_PROGRAM];
 
 export const SidebarWithPrograms = memo(({ items, ...restOfProps }) => {
   const [programs, setPrograms] = useState([]);
@@ -21,7 +25,7 @@ export const SidebarWithPrograms = memo(({ items, ...restOfProps }) => {
   const programsNav = find(itemsWithPrograms, { key: 'programs' });
   if (programs.length > 0) {
     programsNav.hidden = false;
-    programsNav.children = programs.map(({ name, id: id }) => ({
+    programsNav.children = programs.map(({ name, id }) => ({
       label: name,
       path: `/programs/${id}/patients`,
       icon: submenuIcons.action,